backend/requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database
pymongo>=4.6.0

# Web Framework
Flask>=3.0.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0

# Machine Learning & NLP
scikit-learn>=1.3.0
nltk>=3.8.0
textblob>=0.17.1

# Word Cloud Generation
wordcloud>=1.9.0
Pillow>=10.0.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0

# Data validation
jsonschema>=4.20.0